Move context get_metrics/create_metrics to metrics mod
