+++ nucliadb_core/src/lib.rs
@@ -18,7 +18,6 @@
 // along with this program. If not, see <http://www.gnu.org/licenses/>.
 //
 
-pub mod context;
 pub mod fs_state;
 pub mod metrics;
 pub mod paragraphs;
