+++ nucliadb_node/src/http_server/metrics_service.rs
@@ -18,10 +18,10 @@
 // along with this program. If not, see <http://www.gnu.org/licenses/>.
 //
 
-use nucliadb_core::{context, tracing};
+use nucliadb_core::{metrics, tracing};
 
 pub async fn metrics_service() -> String {
-    let metrics = context::get_metrics();
+    let metrics = metrics::get_metrics();
     match metrics.export() {
         Ok(m) => m,
         Err(err) => {
