+++ nucliadb_vectors/src/service/writer.rs
@@ -21,7 +21,7 @@ use std::fmt::Debug;
 use std::time::SystemTime;
 
 use data_point::{Elem, LabelDictionary};
-use nucliadb_core::context;
+use nucliadb_core::metrics;
 use nucliadb_core::metrics::request_time;
 use nucliadb_core::prelude::*;
 use nucliadb_core::protos::prost::Message;
@@ -61,7 +61,7 @@ impl VectorWriter for VectorWriterService {
         let indexset_slock = self.indexset.get_slock()?;
         self.indexset.index_keys(&mut collector, &indexset_slock);
 
-        let metrics = context::get_metrics();
+        let metrics = metrics::get_metrics();
         let took = time.elapsed().map(|i| i.as_secs_f64()).unwrap_or(f64::NAN);
         let metric = request_time::RequestTimeKey::vectors("list_vectorsets".to_string());
         metrics.record_request_time(metric, took);
@@ -86,7 +86,7 @@ impl VectorWriter for VectorWriterService {
             .get_or_create(indexid, similarity, &indexset_elock)?;
         self.indexset.commit(indexset_elock)?;
 
-        let metrics = context::get_metrics();
+        let metrics = metrics::get_metrics();
         let took = time.elapsed().map(|i| i.as_secs_f64()).unwrap_or(f64::NAN);
         let metric = request_time::RequestTimeKey::vectors("add_vectorset".to_string());
         metrics.record_request_time(metric, took);
@@ -105,7 +105,7 @@ impl VectorWriter for VectorWriterService {
         self.indexset.remove_index(indexid, &indexset_elock)?;
         self.indexset.commit(indexset_elock)?;
 
-        let metrics = context::get_metrics();
+        let metrics = metrics::get_metrics();
         let took = time.elapsed().map(|i| i.as_secs_f64()).unwrap_or(f64::NAN);
         let metric = request_time::RequestTimeKey::vectors("remove_vectorset".to_string());
         metrics.record_request_time(metric, took);
@@ -129,7 +129,7 @@ impl WriterChild for VectorWriterService {
         let no_nodes = self.index.no_nodes(&lock);
         std::mem::drop(lock);
 
-        let metrics = context::get_metrics();
+        let metrics = metrics::get_metrics();
         let took = time.elapsed().map(|i| i.as_secs_f64()).unwrap_or(f64::NAN);
         let metric = request_time::RequestTimeKey::vectors("count".to_string());
         metrics.record_request_time(metric, took);
@@ -147,7 +147,7 @@ impl WriterChild for VectorWriterService {
         self.index.delete(&resource_id.uuid, temporal_mark, &lock);
         self.index.commit(lock)?;
 
-        let metrics = context::get_metrics();
+        let metrics = metrics::get_metrics();
         let took = time.elapsed().map(|i| i.as_secs_f64()).unwrap_or(f64::NAN);
         let metric = request_time::RequestTimeKey::vectors("delete_resource".to_string());
         metrics.record_request_time(metric, took);
@@ -301,7 +301,7 @@ impl WriterChild for VectorWriterService {
             debug!("{id:?} - Creating and geting indexes in the set: ends {v} ms");
         }
 
-        let metrics = context::get_metrics();
+        let metrics = metrics::get_metrics();
         let took = time.elapsed().map(|i| i.as_secs_f64()).unwrap_or(f64::NAN);
         let metric = request_time::RequestTimeKey::vectors("set_resource".to_string());
         metrics.record_request_time(metric, took);
@@ -324,7 +324,7 @@ impl WriterChild for VectorWriterService {
             self.collect_garbage_for(&index)?;
         }
 
-        let metrics = context::get_metrics();
+        let metrics = metrics::get_metrics();
         let took = time.elapsed().map(|i| i.as_secs_f64()).unwrap_or(f64::NAN);
         let metric = request_time::RequestTimeKey::vectors("garbage_collection".to_string());
         metrics.record_request_time(metric, took);
